fix(lint): split the 3 files that had grown past the 700-line pre-push gate
`src/routines/service.rs` (701), `src/routines/cleanup/cleanup_tests.rs`
(740), and `src/service/mod_tests.rs` (816) had all crept past the repo's
700-line pre-push gate, which currently fails on plain main for anyone who
has the git hooks installed per CONTRIBUTING.md. Splits each into a sibling
file following the repo's existing per-concern module convention. No
behavior or test-body changes; the full pre-push gate now exits 0 on main.
